requirements.txt
allure_python_commons
colorlog~=6.6.0
openpyxl~=3.0.9
Pillow~=10.4.0
Pillow
PyMySQL~=1.0.2
pytesseract~=0.3.10
pytest==6.2.4
PyYAML~=6.0.1
PyYAML==6.0.1
Requests==2.31.0
selenium==4.16.0

Faker~=9.8.3
redis~=4.3.6